feat(viz): add a gui app in viz app
Current state is that I can display tasks with status bar for each one.
Scrolling vertically (via scroll area) and horizontally (bespoke code,
no scroll bar) is implemented and the task labels stick to the left side
as I want.
